requirements.txt
pydantic>=2.0.0
openai>=1.0.0
python-dotenv
streamlit